Moving TestAccessors to separate files.
